third_party/WebKit/LayoutTests/fast/dom/constructor-proto.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>

description('This tests that setting the proto of a built in constructor is not reset');

shouldBe('Object.getPrototypeOf(HTMLElement)', 'Element');

var proto = new Number(42);
HTMLElement.__proto__ = proto;
shouldBe('Object.getPrototypeOf(HTMLElement)', 'proto');

</script>
